Extract model processing to its own file
